refactoring to convert into a python module
