+++ scripts/.gitignore
@@ -1,2 +1,2 @@
-temp_CPSL_Radar
+temp_RadCloud
 *.mp4
