+++ scripts/generate_dataset_radar_only.py
@@ -5,7 +5,7 @@ import pandas as pd
 import matplotlib.pyplot as plt
 
 #import CPSL Radar Dataset Generator Code
-from CPSL_Radar.datasets.Dataset_Generator import DatasetGenerator
+from RadCloud.datasets.Dataset_Generator import DatasetGenerator
 
 drone_folder = "/home/david/CPSL_Drone/"
 # test_scenarios = ["drone_test"]
